Separated the parts field in the jobpart rows into different fields for different part statuses

+++ backend/GarageControl.Core/Contracts/INotificationService.cs
@@ -7,6 +7,6 @@ namespace GarageControl.Core.Contracts
         Task<List<NotificationViewModel>> GetUserNotificationsAsync(string userId);
         Task MarkAsReadAsync(string notificationId, string userId);
         Task DeleteOldNotificationsAsync();
-        Task SendStockNotificationAsync(string workshopId, string partId, string partName, int currentBalance, int minQuantity);
+        Task SendStockNotificationAsync(string workshopId, string partId, string partName, double currentBalance, double minQuantity);
     }
 }
